+++ CHANGELOG.md
@@ -5,7 +5,7 @@ file.
 The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [2.0.0] - 2023-02-08
 ### Added
 - `credentials.json` file for storing Monkey Island user login information. #1206
 - `GET /api/propagation-credentials/<string:guid>` endpoint for agents to
@@ -159,6 +159,11 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - A bug where the Log4Shell exploiter could leave LDAP servers and child
   processes running. #2820
 - A bug in registration process that caused the button to be stuck with loading icon. #2916
+- Configurability of SSH key pairs. #1416
+- A bug in the security report that didn't show the correct percentage of exploited machines. #2954
+- A bug where ransomware README file is not readable on older Windows machines. #2951
+- An exception being raised if the ransomware target directory does not exist. #2953
+- A bug where the ransomware payload could follow a symlink. #2953
 
 ### Security
 - Upgrade Cryptography dependency. #1482